+++ changelog.md
@@ -1,6 +1,6 @@
 ### Changelog next version
 
-* Improved `SpringMongoEventStore` so that it never does in-memory filtering of events that we're not interested in.
+* Improved `SpringMongoEventStore`, `MongoEventStore` so that it never does in-memory filtering of events that we're not interested in.
 
 ### Changelog 0.14.1 (2021-11-12)