+++ changelog.md
@@ -1,6 +1,7 @@
 ### Changelog next version
 
 * Improved `SpringMongoEventStore`, `MongoEventStore` and `ReactorMongoEventStore` so that it never does in-memory filtering of events that we're not interested in.
+* Upgraded to Kotlin 1.6.0
 
 ### Changelog 0.14.1 (2021-11-12)